Decouple "aesgcm" and "aes128gcm" schemes, restruture record chunking for "aes128gcm".
This is a significant refactor of the guts of the crypto code, but I think overall it
makes things easier to understand and to audit.

First, I've removed the `EceWebPush` trait that was previously used to share parts
of the encrypt/decrypt logic between the two schemes. The schemes are not that similar
in practice and on balance, I think the attempt to share code between them was
actually making both schemes harder to understand.

Second, I've cut all the record-chunking code out of "aesgcm". It now supports only
a single record on both encryption and decryption, in line with what the spec says
that a webpush client should support. We were already throwing errors when encountering
multiple records in "aesgcm"; this cleanup takes advantage of that fact to actually
remove the code without breaking the public API.

Finally, I've re-worked the "aes128gcm" implementation to try to make its operation
easier to understand. I've added several layers of abstraction in the hope that
readers can process them one at a time - for example there is a separate `Header`
struct for reading/writing the header, there is a separate iterator that just does
the chunking into fixed-size records, etc.

Of course "easier to understand" is subjective, but I think it's an improvement
(and I certainly understand things better as a result of having worked through it!).
Feedback and/or pushback on this is most welcome.

Fixes #55.